split out ComboInfoPanelComponent;
